requirements.txt
requests
selenium
PyYAML